refactor(api): deduplicate workflow controller schemas into controller_schemas.py
